Now accepts filenames or streams as input.
If input parameter has a read method, it is called expecting to get a
String which is then used as input.  Otherwise if the input is a
kind_of? String it is treated as a filename and that file is opened
and read as the input.

Otherwise an exception is raised.